\name{shiny.app}
\alias{shiny.app}
\title{Run Shiny application}
\usage{
  shiny.app()
}
\description{
  Runs a Swiss Mortgage Calculator Shiny application. This
  function normally does not return; interrupt R to stop
  the application (usually by pressing Ctrl+C or Esc).
}
\examples{
\dontrun{
shiny.app()
}
}

